# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SecuritySuite (secapi) — APIs propias",
    "version": "0.1.0",
    "description": "Catálogo generado por `pnpm docs:api` a partir de `src/app/api/**/route.ts`.\n\nLo generado describe **lo que el código hace hoy**, incluida la autenticación real de\ncada handler. Las descripciones a mano viven en `docs/api/anotaciones.yaml` y ganan\nsobre lo inferido; `GET /api/docs/spec` devuelve el documento ya mergeado.\n\nNo editar este archivo a mano: se regenera."
  },
  "servers": [
    {
      "url": "http://localhost:4005",
      "description": "desarrollo local (pnpm dev)"
    },
    {
      "url": "https://secapi.riogas.com.uy",
      "description": "secapi (PM2, puerto 3001)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerJwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT emitido por `POST /api/db/login`. Payload: `{ iss, username, userId, sistema }` — no lleva el flag root."
      },
      "cookieToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "token",
        "description": "El mismo JWT, tomado de la cookie `token` cuando no hay header."
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
  "paths": {
...
    "/api/db/aplicaciones": {
      "get": {
        "tags": [
          "aplicaciones"
        ],
        "summary": "GET /api/db/aplicaciones?filtro=&estado=&page=1&pageSize=20",
        "description": "GET /api/db/aplicaciones?filtro=&estado=&page=1&pageSize=20",
        "operationId": "get_api_db_aplicaciones",
        "parameters": [
          {
            "name": "estado",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Detectado estáticamente (`searchParams.get`); anotalo para describirlo."
          },
          {
            "name": "filtro",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Detectado estáticamente (`searchParams.get`); anotalo para describirlo."
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Detectado estáticamente (`searchParams.get`); anotalo para describirlo."
          },
          {
            "name": "pageSize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Detectado estáticamente (`searchParams.get`); anotalo para describirlo."
          }
        ],
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/aplicaciones/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      },
      "post": {
        "tags": [
          "aplicaciones"
        ],
        "summary": "POST /api/db/aplicaciones — crear",
        "description": "POST /api/db/aplicaciones — crear",
        "operationId": "post_api_db_aplicaciones",
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/aplicaciones/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      }
    },
...
    "/api/db/funcionalidades/{id}": {
      "get": {
        "tags": [
          "funcionalidades"
        ],
        "summary": "GET /api/db/funcionalidades/[id]",
        "description": "GET /api/db/funcionalidades/[id]",
        "operationId": "get_api_db_funcionalidades_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Segmento dinámico `[id]` de la ruta."
          }
        ],
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/funcionalidades/[id]/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      },
      "put": {
        "tags": [
          "funcionalidades"
        ],
        "summary": "PUT /api/db/funcionalidades/[id]",
        "description": "PUT /api/db/funcionalidades/[id]",
        "operationId": "put_api_db_funcionalidades_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Segmento dinámico `[id]` de la ruta."
          }
        ],
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/funcionalidades/[id]/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      },
      "delete": {
        "tags": [
          "funcionalidades"
        ],
        "summary": "DELETE /api/db/funcionalidades/[id] — soft delete",
        "description": "DELETE /api/db/funcionalidades/[id] — soft delete",
        "operationId": "delete_api_db_funcionalidades_id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Segmento dinámico `[id]` de la ruta."
          }
        ],
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/funcionalidades/[id]/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      }
    },
    "/api/db/funcionalidades/{id}/acciones": {
      "get": {
        "tags": [
          "funcionalidades"
        ],
        "summary": "Devuelve los ObjetoAccion seleccionados para esta funcionalidad",
        "description": "GET /api/db/funcionalidades/[id]/acciones\nDevuelve los ObjetoAccion seleccionados para esta funcionalidad",
        "operationId": "get_api_db_funcionalidades_id_acciones",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Segmento dinámico `[id]` de la ruta."
          }
        ],
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/funcionalidades/[id]/acciones/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      },
      "put": {
        "tags": [
          "funcionalidades"
        ],
        "summary": "Reemplaza el set completo de items seleccionados",
        "description": "PUT /api/db/funcionalidades/[id]/acciones\nReemplaza el set completo de items seleccionados\nBody: { items: Array<{ objetoId: number, objetoAccionId?: number | null }> }",
        "operationId": "put_api_db_funcionalidades_id_acciones",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Segmento dinámico `[id]` de la ruta."
          }
        ],
        "responses": {
          "default": {
            "description": "Sin anotar todavía (ver docs/api/README.md)."
          }
        },
        "x-auth": {
          "modo": "jwt",
          "verificaFirma": true,
          "detectadoEn": "handler",
          "detalle": "JWT en `Authorization: Bearer` o cookie `token`, con firma y vencimiento verificados por el guard central de `/api/db` (`requireApiAuth`, src/lib/auth/apiGuard.ts). El nivel exigido por esta ruta está declarado en `POLITICAS`, en ese mismo archivo."
        },
        "x-archivo": "src/app/api/db/funcionalidades/[id]/acciones/route.ts",
        "x-generado": true,
        "security": [
          {
            "bearerJwt": []
          },
          {
            "cookieToken": []
          }
        ]
      }
    },
...
  },
...
}